use size_t in audioOutput.c internals
git-svn-id: https://svn.musicpd.org/mpd/trunk@7294 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/audioOutput.c
@@ -188,7 +188,7 @@ static void convertAudioFormat(AudioOutput * audioOutput,
 					   *sizeArgPtr,
 					   &(audioOutput->outAudioFormat));
 
-	if (size > (size_t)audioOutput->convBufferLen) {
+	if (size > audioOutput->convBufferLen) {
 		audioOutput->convBuffer =
 		    xrealloc(audioOutput->convBuffer, size);
 		audioOutput->convBufferLen = size;
